Implement support for --output, -o <file>
This change refactors tool's option handling to support "const" options,
which:

   - can be set only once
   - canont perform any action

Because of those properties, "const" options can be processed before all
the others. With that mechanism, now we process the options twice: once
with the `const_processing` flag set, at the beginning in main() and later
when performing the actions.

That "fancy" options handling allows the tool to accept a series of actions
to perform in the given order using options which can change multiple times
between the actions. Now with the "const" options, we have also classic
options which can never change and their position is irrelevant, so they
can be processed separately.

The --output, -o option is a "const" option and it couldn't be implemented
with the previous elfhack option handling logic simply because we first
open the ELF binary, memory map it and only then we start processing the
options. While for the output flag, we clearly need that information much
earlier.